+++ docs/changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.5.3   (unreleased)
+1.5.3   (2017-11-13)
 --------------------
 
 Modifications to existing flavors: